+++ common/src/main/java/org/adde0109/pcf/forwarding/PacketDecoder.java
@@ -14,7 +14,6 @@ import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.MessageToMessageDecoder;
 
 import org.adde0109.pcf.PCF;
-import org.adde0109.pcf.forwarding.modern.ServerLoginPacketListenerBridge;
 import org.jspecify.annotations.NonNull;
 
 import java.util.List;
