+++ deobsf/v26_1/src/main/resources/pcf.mixins.v26_1.json
@@ -11,9 +11,9 @@
     "v26_1.crossstich.ArgumentNodeStubMixin",
     "v26_1.crossstich.ArgumentTypeInfoMixin",
     "v26_1.forwarding.ConnectionMixin",
+    "v26_1.forwarding.ServerLoginPacketListenerImplMixin",
     "v26_1.forwarding.legacy.ServerLoginPacketListenerImplHelloMixin",
     "v26_1.forwarding.modern.ServerLoginPacketListenerImplHelloHybridMixin",
-    "v26_1.forwarding.modern.ServerLoginPacketListenerImplHelloMixin",
-    "v26_1.forwarding.modern.ServerLoginPacketListenerImplMixin"
+    "v26_1.forwarding.modern.ServerLoginPacketListenerImplHelloMixin"
   ]
 }
