+++ legacy/v12_2/src/forge/resources/pcf.mixins.v12_2.forge.json
@@ -9,8 +9,8 @@
   "package": "org.adde0109.pcf.mixin",
   "server": [
     "v12_2.forge.forwarding.ConnectionMixin",
-    "v12_2.forge.forwarding.modern.ServerLoginPacketListenerImplHelloMixin",
-    "v12_2.forge.forwarding.modern.ServerLoginPacketListenerImplMixin",
-    "v12_2.forge.forwarding.modern.ServerLoginPacketListenerImplMixin$SLPLIMixin_12"
+    "v12_2.forge.forwarding.ServerLoginPacketListenerImplMixin",
+    "v12_2.forge.forwarding.ServerLoginPacketListenerImplMixin$SLPLIMixin_12",
+    "v12_2.forge.forwarding.modern.ServerLoginPacketListenerImplHelloMixin"
   ]
 }
