+++ modern/v13_2/src/forge/java/org/adde0109/pcf/mixin/v13_2/forge/forwarding/ServerLoginPacketListenerImplMixin.java
@@ -1,4 +1,4 @@
-package org.adde0109.pcf.mixin.v13_2.forge.forwarding.modern;
+package org.adde0109.pcf.mixin.v13_2.forge.forwarding;
 
 import static org.adde0109.pcf.forwarding.modern.ModernForwarding.handleHello;
 
@@ -14,7 +14,7 @@ import net.minecraft.network.NetworkManager;
 import net.minecraft.util.text.ITextComponent;
 
 import org.adde0109.pcf.forwarding.ConnectionBridge;
-import org.adde0109.pcf.forwarding.modern.ServerLoginPacketListenerBridge;
+import org.adde0109.pcf.forwarding.ServerLoginPacketListenerBridge;
 import org.apache.logging.log4j.Logger;
 import org.jspecify.annotations.NonNull;
 import org.jspecify.annotations.Nullable;
