+++ modern/v13_2/src/forge/resources/pcf.mixins.v13_2.forge.json
@@ -9,6 +9,6 @@
   "package": "org.adde0109.pcf.mixin",
   "server": [
     "v13_2.forge.crossstitch.CommandsPacketMixin",
-    "v13_2.forge.forwarding.modern.ServerLoginPacketListenerImplMixin"
+    "v13_2.forge.forwarding.ServerLoginPacketListenerImplMixin"
   ]
 }
