+++ modern/v16_5/src/forge/java/org/adde0109/pcf/mixin/v16_5/forge/forwarding/modern/ServerLoginPacketListenerImplHelloMixin.java
@@ -2,26 +2,18 @@ package org.adde0109.pcf.mixin.v16_5.forge.forwarding.modern;
 
 import static org.adde0109.pcf.forwarding.modern.ModernForwarding.handleHello;
 
-import com.mojang.authlib.GameProfile;
-
 import dev.neuralnexus.taterapi.meta.Mappings;
 import dev.neuralnexus.taterapi.meta.anno.AConstraint;
 import dev.neuralnexus.taterapi.meta.anno.Versions;
 import dev.neuralnexus.taterapi.meta.enums.MinecraftVersion;
 import dev.neuralnexus.taterapi.meta.enums.Platform;
 
-import net.minecraft.network.Connection;
-import net.minecraft.network.chat.Component;
 import net.minecraft.server.network.ServerLoginPacketListenerImpl;
 
-import org.adde0109.pcf.forwarding.ConnectionBridge;
-import org.adde0109.pcf.forwarding.modern.ServerLoginPacketListenerBridge;
+import org.adde0109.pcf.forwarding.ServerLoginPacketListenerBridge;
 import org.apache.commons.lang3.Validate;
-import org.apache.logging.log4j.Logger;
 import org.jspecify.annotations.NonNull;
-import org.jspecify.annotations.Nullable;
 import org.objectweb.asm.Opcodes;
-import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
@@ -32,17 +24,11 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
         mappings = Mappings.SEARGE,
         version = @Versions(min = MinecraftVersion.V14, max = MinecraftVersion.V16_5))
 @Mixin(ServerLoginPacketListenerImpl.class)
-public abstract class ServerLoginPacketListenerImplMixin
+public abstract class ServerLoginPacketListenerImplHelloMixin
         implements ServerLoginPacketListenerBridge {
     // spotless:off
-    @Shadow @Final public Connection connection;
-    @Shadow @Nullable private GameProfile gameProfile;
     @Shadow private ServerLoginPacketListenerImpl.State state;
-    @Shadow @Final private static Logger LOGGER;
-    @Shadow public abstract void shadow$disconnect(Component reason);
-    // spotless:on
 
-    // spotless:off
     @AConstraint(
             platform = {Platform.ARCLIGHT, Platform.CATSERVER, Platform.MAGMA, Platform.MOHIST}, invert = true)
     @Inject(method = "handleHello", cancellable = true, at = @At(value = "FIELD", opcode = Opcodes.PUTFIELD, ordinal = 1,
@@ -63,36 +49,5 @@ public abstract class ServerLoginPacketListenerImplMixin
         Validate.validState(
                 this.state == ServerLoginPacketListenerImpl.State.HELLO, "Unexpected hello packet");
         handleHello(this, ci);
-    }
-
-    @Override
-    public @NonNull ConnectionBridge bridge$connection() {
-        return (ConnectionBridge) this.connection;
-    }
-
-    @Override
-    public void bridge$disconnect(final @NonNull Object reason) {
-        this.shadow$disconnect((Component) reason);
-    }
-
-    @Override
-    public void bridge$setGameProfile(final @NonNull GameProfile profile) {
-        this.gameProfile = profile;
-    }
-
-    @Override
-    public void bridge$startClientVerification(final @NonNull GameProfile profile) {
-        this.gameProfile = profile;
-        this.state = ServerLoginPacketListenerImpl.State.NEGOTIATING;
-    }
-
-    @Override
-    public void bridge$logger_info(final @NonNull String text, final Object... params) {
-        LOGGER.info(text, params);
-    }
-
-    @Override
-    public void bridge$logger_error(final @NonNull String text, final Object... params) {
-        LOGGER.error(text, params);
     }
 }
