+++ modern/v16_5/src/forge/resources/pcf.mixins.v16_5.forge.json
@@ -12,6 +12,7 @@
     "v16_5.forge.crossstitch.ArgumentTypesEntryMixin",
     "v16_5.forge.crossstitch.CommandsPacketMixin",
     "v16_5.forge.forwarding.ConnectionMixin",
-    "v16_5.forge.forwarding.modern.ServerLoginPacketListenerImplMixin"
+    "v16_5.forge.forwarding.ServerLoginPacketListenerImplMixin",
+    "v16_5.forge.forwarding.modern.ServerLoginPacketListenerImplHelloMixin"
   ]
 }
