+++ modern/v19_2/src/forge/java/org/adde0109/pcf/mixin/v19_2/forge/forwarding/ServerLoginPacketListenerImplStartClientVerificationMixin.java
@@ -1,4 +1,4 @@
-package org.adde0109.pcf.mixin.v19_2.forge.forwarding.modern;
+package org.adde0109.pcf.mixin.v19_2.forge.forwarding;
 
 import com.mojang.authlib.GameProfile;
 
@@ -9,7 +9,7 @@ import dev.neuralnexus.taterapi.meta.enums.MinecraftVersion;
 
 import net.minecraft.server.network.ServerLoginPacketListenerImpl;
 
-import org.adde0109.pcf.forwarding.modern.ServerLoginPacketListenerBridge;
+import org.adde0109.pcf.forwarding.ServerLoginPacketListenerBridge;
 import org.jspecify.annotations.NonNull;
 import org.jspecify.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
